requirements.txt
absl-py==0.13.0
altgraph==0.17
APScheduler==3.6.3
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
comtypes==1.1.10
cvzone==1.5.0
cycler==0.10.0
distlib==0.3.2
Django==3.2.6
djangorestframework==3.12.4
docopt==0.6.2
enum34==1.1.10
ez-setup==0.9
filelock==3.0.12
Flask==2.0.1
future==0.18.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Js2Py==0.71
keyboard==0.13.5
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
mediapipe==0.8.7.2
MouseInfo==0.1.3
msgbox==0.6.0
numpy==1.21.2
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
packaging==21.0
pefile==2021.9.3
Pillow==8.3.2
pipwin==0.5.1
platformdirs==2.2.0
plyer==2.0.0
protobuf==3.17.3
psutil==5.8.0
psycopg2==2.9.1
pyaes==1.6.1
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycaw==20181226
pycodestyle==2.7.0
pydub==0.25.1
PyGetWindow==0.0.9
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyPrind==2.11.3
PyRect==0.1.4
PyScreeze==0.1.27
pySmartDL==1.3.4
python-dateutil==2.8.2
python-telegram-bot==13.7
pyttsx3==2.90
pytube==11.0.0
PyTweening==1.0.3
pytz==2021.1
pyvbox==2.1.1
pywin32==301
pywin32-ctypes==0.2.0
requests==2.26.0
rsa==4.7.2
scipy==1.7.1
six==1.16.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
Telethon==1.23.0
termcolor==1.1.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
virtualbox==2.1.1
virtualenv==20.7.2
Werkzeug==2.0.1
wikipedia==1.4.0
XlsxWriter==3.0.1
youtube-search==2.1.0
